requirements.txt
django>=2.0,<2.1
django-bootstrap4
psutil
python-twitter
jinja2
